+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/MPRPreset.java
@@ -3,12 +3,10 @@ package insane96mcp.mobspropertiesrandomness.data.json;
 import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.annotations.SerializedName;
 import insane96mcp.insanelib.exception.JsonValidationException;
-import insane96mcp.mobspropertiesrandomness.MobsPropertiesRandomness;
+import insane96mcp.mobspropertiesrandomness.data.json.properties.MPRBossBar;
 import insane96mcp.mobspropertiesrandomness.setup.Strings;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.bossevents.CustomBossEvent;
-import net.minecraft.server.bossevents.CustomBossEvents;
-import net.minecraft.world.BossEvent;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.level.Level;
 
@@ -16,27 +14,21 @@ public class MPRPreset extends MPRProperties implements IMPRObject {
 	@JsonAdapter(ResourceLocation.Serializer.class)
 	public transient ResourceLocation id;
 
-	@SerializedName("show_boss_bar")
-	public Boolean showBossBar;
+	@SerializedName("boss_bar")
+	public MPRBossBar bossBar;
 
 	@Override
 	public void validate() throws JsonValidationException {
 		super.validate();
-
-		if (this.showBossBar == null)
-			this.showBossBar = false;
 	}
 
 	@Override
 	public boolean apply(LivingEntity livingEntity, Level level) {
 		boolean ret = super.apply(livingEntity, level);
-		if (this.showBossBar) {
-			CustomBossEvents bossEvents = livingEntity.getServer().getCustomBossEvents();
-			ResourceLocation bossBarId = new ResourceLocation(MobsPropertiesRandomness.MOD_ID, id.getPath() + "_" + livingEntity.getRandom().nextInt(Integer.MAX_VALUE));
-			CustomBossEvent bossEvent = bossEvents.create(bossBarId, livingEntity.getDisplayName());
-			bossEvent.setColor(BossEvent.BossBarColor.PINK);
-			bossEvent.setOverlay(BossEvent.BossBarOverlay.NOTCHED_10);
-			livingEntity.getPersistentData().putString(Strings.Tags.BOSS_BAR_ID, bossBarId.toString());
+		if (this.bossBar != null) {
+			//noinspection ConstantConditions
+			CustomBossEvent bossBar = this.bossBar.createBar(this.id.getPath(), livingEntity, livingEntity.getRandom(), livingEntity.getServer());
+			livingEntity.getPersistentData().putString(Strings.Tags.BOSS_BAR_ID, bossBar.getTextId().toString());
 		}
 		return ret;
 	}